Ajout confiormation suppression groupe+correction bug filtres

+++ GestAssoc/GestAssoc.Modules.GestionGroupes/Commands/DeleteGroupeCommand.cs
@@ -1,9 +1,8 @@
-﻿using GestAssoc.Common.Commands;
-using GestAssoc.Common.Utility;
+﻿using GestAssoc.Common.Utility;
 using GestAssoc.Model.Models;
-using GestAssoc.Modules.GestionGroupes.Constantes;
 using GestAssoc.Modules.GestionGroupes.Properties;
 using GestAssoc.Modules.GestionGroupes.Services;
+using Microsoft.Practices.Prism.Interactivity.InteractionRequest;
 using Microsoft.Practices.ServiceLocation;
 using Microsoft.Practices.Unity;
 using System;
@@ -13,6 +12,14 @@ namespace GestAssoc.Modules.GestionGroupes.Commands
 {
 	public class DeleteGroupeCommand : ICommand
 	{
+		public InteractionRequest<IConfirmation> RqConfirmDelete { get; private set; }
+		private Action _commandCallBack;
+
+		public DeleteGroupeCommand(Action callback) {
+			this.RqConfirmDelete = new InteractionRequest<IConfirmation>();
+			this._commandCallBack = callback;
+		}
+		
 		public bool CanExecute(object parameter) {
 			var itemToDelete = parameter as Groupe;
 
@@ -26,22 +33,33 @@ namespace GestAssoc.Modules.GestionGroupes.Commands
 		}
 
 		public void Execute(object parameter) {
-			var service = ServiceLocator
-				.Current.GetInstance<IUnityContainer>()
-				.Resolve<IGestionGroupesServices>();
+			this.RqConfirmDelete.Raise(
+				new Confirmation
+				{
+					Content = Resources.Confirm_SuppressionGroupe + Environment.NewLine + (parameter as Groupe).ToString(),
+					Title = Common.Properties.Resources.Titre_Confirmation
+				},
+				c => this.ExecuteCallback(c.Confirmed, parameter as Groupe)
+			);
+		}
 
-			var itemToDelete = parameter as Groupe;
+		private void ExecuteCallback(bool deleteConfirmed, Groupe itemToDelete) {
+			if (deleteConfirmed) {
+				var service = ServiceLocator
+					.Current.GetInstance<IUnityContainer>()
+					.Resolve<IGestionGroupesServices>();
 
-			try {
-				UIServices.SetBusyState();
-				service.DeleteGroupe(itemToDelete);
+				try {
+					UIServices.SetBusyState();
+					service.DeleteGroupe(itemToDelete);
 
-				NotificationHelper.WriteLog(Resources.Log_EnregistrementSupprime);
+					NotificationHelper.WriteLog(Resources.Log_EnregistrementSupprime);
 
-				new ShowViewCommand(ViewNames.ConsultationGroupes.ToString()).Execute(null);
-			}
-			catch (Exception ex) {
-				NotificationHelper.ShowError(ex);
+					this._commandCallBack();
+				}
+				catch (Exception ex) {
+					NotificationHelper.ShowError(ex);
+				}
 			}
 		}
 	}
